Broke out 'Tutorial' into own class, Work on calling it in Game
